Substantial rework of grapheme and grapheme_view; add {text,rope}::insert(grapheme); add code using this insert() overload to the case mapping example.

+++ include/boost/text/case_mapping.hpp
@@ -587,7 +587,7 @@ namespace boost { namespace text {
                 if (special_dutch_case &&
                     title_state == title_state_t::right_after &&
                     (*it == 'j' || *it == 'J')) {
-                    *out = 'J';
+                    *out = uint32_t('J');
                     ++out;
                     ++it;
                     continue;
